+++ src/main/java/twenty/day/fourteen/MemoryInstruction.java
@@ -1,4 +1,4 @@
 package twenty.day.fourteen;
 
-record MemoryInstruction(Integer memoryAddress, Integer value) {
+record MemoryInstruction(Integer memoryAddress, Long value) {
 }
